feat: Implement initial core modules for Jasa Angkut, Bengkel, SDM, and Keuangan management across frontend and backend.

+++ frontend/app/sdm/slip-gaji.tsx
@@ -171,7 +171,7 @@ export default function SlipGajiScreen() {
     }, [refetchList, refetchSummary, refetchPreview]);
 
     const handleUpdateAttendance = (karyawanId: number, value: string) => {
-        const numValue = parseInt(value) || 0;
+        const numValue = parseFloat(value) || 0;
         setAttendanceEdits(prev => ({
             ...prev,
             [karyawanId]: numValue
